% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideology.R
\name{add_station_dummies}
\alias{add_station_dummies}
\title{add_station_dummies}
\usage{
add_station_dummies(.tbl)
}
\description{
Adds two dummies: is_private captures private stations (vs public ones) and is_tv captures TV stations (vs radio ones)
}
